backend/requirements.txt
streamlit
Pillow
requests
pymupdf
pdf2image
aiohttp
azure-ai-vision-imageanalysis
azure-ai-documentintelligence
azure-core
tiktoken
python-dotenv
evaluate
jiwer
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy>=1.26.4
pandas>=2.0.0
openpyxl>=3.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.2.30
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
azure-storage-blob 
azure-identity
pydantic[email]
celery>=5.3.0
redis>=5.0.0
nest-asyncio>=1.6.0
reportlab>=4.0.0